+++ client/styles.js
@@ -26,6 +26,7 @@ import '/client/components/import/import.css';
 import '/client/components/lists/list.css';
 import '/client/components/main/accessibility.css';
 import '/client/components/main/brokenCards.css';
+import '/client/components/main/customTheme.css';
 import '/client/components/main/dueCards.css';
 import '/client/components/main/editor.css';
 import '/client/components/main/fonts.css';